Update code (database.py, habit_tracker.py, habit.py)

+++ habit_tracker.py
@@ -1,12 +1,12 @@
 from database import HabitDatabase
-
+from habit import Habit 
 
 class HabitTracker:
     """Tracks habits and their states."""
 
-    def __init__(self):
+    def __init__(self, habit_database: HabitDatabase):
         """Initialize HabitTracker."""
-        self.habit_database = HabitDatabase()
+        self.habit_database = habit_database
         self.habits = self.habit_database.get_habits()
 
     def add_habit(self, habit):
@@ -15,8 +15,11 @@ class HabitTracker:
 
         :param habit: Habit object to add to the tracker
         """
-        self.habit_database.add_habit(habit)
-        self.habits = self.habit_database.get_habits()
+        if habit.name not in self.habit_database.get_habits():
+            self.habit_database.add_habit(habit.name, habit.task_specification)
+            self.habits.append(habit)
+        else:
+            raise ValueError(f"Habit with name '{habit.name}' already exists.")
 
     def remove_habit(self, habit_name):
         """
@@ -24,8 +27,12 @@ class HabitTracker:
 
         :param habit_name: Name of the habit to remove
         """
-        self.habit_database.remove_habit(habit_name)
-        self.habits = self.habit_database.get_habits()
+        if habit_name in self.habit_database.get_habits():
+            self.habit_database.remove_habit(habit_name)
+            self.habits = self.habit_database.get_habits()  
+            print(f"Habit {habit_name} removed successfully.")
+        else:
+            raise ValueError(f"No habit with name '{habit_name}' found.")
 
     def get_habits(self):
         """
@@ -42,8 +49,7 @@ class HabitTracker:
         :param habit_name: Name of the habit to check
         :return: State of the habit
         """
-        habit_data = self.habit_database.get_habit_data(habit_name)
-        if habit_data:
+        if habit_name in self.habit_database.get_habits():
             return "Habit exists"
         else:
             return "No habits tracked"
@@ -55,4 +61,8 @@ class HabitTracker:
 
     def update_database(self):
         """Update the database with current habit data."""
-        self.habit_database.update_data()
+        self.habit_database.update_database()
+
+    def close_database(self):
+        """Close the connection to the database."""
+        self.habit_database.close()
